fix(#150): AssignmentCard 선생, 학생 분리
