+++ src/features/Common/Class/api.ts
@@ -1,5 +1,5 @@
 import Customapi from '@/shared/config/api';
-import { Assignment } from '@/shared/types/Class/Assignment/assignmentAttachment';
+import { Assignment } from '@/shared/types/Class/Assignment/Attachment';
 import { Exam } from '@/shared/types/Class/Exam';
 import { NewsItem, QuestionItem, Directory } from '@/shared/types/Class/Lesson';
